handle validator iteration inside the validators file
